factor type_parser
